refactor storage to move every struct in its own file
